Split parser package into topdown and bottomup
